Introduce Windows build of Node Problem Detector
